requirements.txt
pyyaml
pytz
python-telegram-bot==20.3
schedule==1.2.0
requests==2.31.0
